[FIX] point service 통합테스트 동시성 테스트 분리
