+++ src/test/java/kr/hhplus/be/server/domain/point/PointServiceIntegrationTest.kt
@@ -5,14 +5,9 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.Executors
-import java.util.concurrent.atomic.AtomicInteger
 
 @SpringBootTest
-class PointServiceIntegrationTest {
-    @Autowired
-    private lateinit var pointRepository: PointRepository
+class PointServiceIntegrationTests {
 
     @Autowired
     private lateinit var pointService: PointService
@@ -47,39 +42,6 @@ class PointServiceIntegrationTest {
             assertThat(result.userId).isEqualTo(userId)
             assertThat(result.balance).isEqualTo(expected.balance + amount)
         }
-
-        @Test
-        fun `happy - 포인트 충전 동시성 테스트`() {
-            // given
-            val userId = 4L
-            val amount = 100L
-            val count = 10
-            val service = Executors.newFixedThreadPool(5)
-            val latch = CountDownLatch(count)
-            val expected = pointService.find(PointCommand.Find(userId))
-            val success = AtomicInteger(0)
-            val failure = AtomicInteger(0)
-            // when
-            repeat(count) {
-                service.submit {
-                    try {
-                        pointService.charge(PointCommand.Charge(userId, amount))
-                        success.incrementAndGet()
-                    } catch (e: Exception) {
-                        failure.incrementAndGet()
-                    } finally {
-                        latch.countDown()
-                    }
-                }
-            }
-            latch.await()
-            service.shutdown()
-            val result = pointService.find(PointCommand.Find(userId))
-            // then
-            assertThat(success.get()).isGreaterThanOrEqualTo(count)
-            assertThat(failure.get()).isLessThanOrEqualTo(0)
-            assertThat(result.balance).isEqualTo(expected.balance + amount * success.get())
-        }
     }
 
     @Nested
@@ -96,39 +58,6 @@ class PointServiceIntegrationTest {
             // then
             assertThat(result.userId).isEqualTo(userId)
             assertThat(result.balance).isEqualTo(expected.balance - amount)
-        }
-
-        @Test
-        fun `happy - 포인트 사용 동시성 테스트`() {
-            // given
-            val userId = 5L
-            val amount = 100L
-            val count = 10
-            val service = Executors.newFixedThreadPool(5)
-            val latch = CountDownLatch(count)
-            val expected = pointService.find(PointCommand.Find(userId))
-            val success = AtomicInteger(0)
-            val failure = AtomicInteger(0)
-            // when
-            repeat(count) {
-                service.submit {
-                    try {
-                        pointService.use(PointCommand.Use(userId, amount))
-                        success.incrementAndGet()
-                    } catch (e: Exception) {
-                        failure.incrementAndGet()
-                    } finally {
-                        latch.countDown()
-                    }
-                }
-            }
-            latch.await()
-            service.shutdown()
-            val result = pointService.find(PointCommand.Find(userId))
-            // then
-            assertThat(success.get()).isGreaterThanOrEqualTo(count)
-            assertThat(failure.get()).isLessThanOrEqualTo(0)
-            assertThat(result.balance).isEqualTo(expected.balance - amount * success.get())
         }
     }
 }
